fix(decisions): cast role member counts to int to fix output validation failure
The drizzle rewrite of the per-role member-count subquery
(c683b09b4) dropped the `::int` cast the original raw SQL relied on.
Postgres's `COUNT(*)` returns `bigint`, and node-postgres decodes
`bigint` as a string rather than a number to avoid silent precision
loss — so without the cast, `profile.listRoles` started returning
`memberCount` as a string instead of a number.

That broke the endpoint's own output contract: the response fails its
Zod number check with a TRPCError ("Output validation failed"),
which surfaced as a CI failure on #1721 once the test suite exercised
`listRoles` against a real Postgres instance rather than a mock.

Re-adds the cast so the count round-trips as a JS number, matching
what the raw-SQL version already did before the rewrite.

+++ packages/common/src/services/access/getRoles.ts
@@ -152,7 +152,9 @@ export const getRoles = async (params?: {
         description: accessRoles.description,
         permission: accessRolePermissionsOnAccessZones.permission,
         ...(shouldIncludeMemberCounts && {
-          memberCount: sql<number>`(${memberCountSubquery(accessRoles)})`,
+          // Postgres COUNT(*) returns bigint, which node-postgres decodes as
+          // a string; cast to int so the value round-trips as a JS number.
+          memberCount: sql<number>`(${memberCountSubquery(accessRoles)})::int`,
         }),
       })
       .from(accessRoles)
@@ -215,7 +217,9 @@ export const getRoles = async (params?: {
     extras: {
       memberCount: (table, { sql: sqlOp }) =>
         shouldIncludeMemberCounts
-          ? sqlOp<number>`(${memberCountSubquery(table)})`.as('member_count')
+          ? sqlOp<number>`(${memberCountSubquery(table)})::int`.as(
+              'member_count',
+            )
           : sqlOp<number>`0`.as('member_count'),
     },
     limit: limit + 1,
